requirements.txt
streamlit>=1.35.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
piexif>=1.1.3
Pillow>=10.0.0
pandas>=2.0.0
CairoSVG>=2.7.0
